requirements.txt
transformers>=4.41.0
datasets>=2.19.0
peft>=0.11.0
trl>=0.9.0
accelerate>=0.30.0
bitsandbytes>=0.43.0
torch>=2.2.0
pandas>=2.2.0
matplotlib>=3.8.0
pyyaml>=6.0.1
tqdm>=4.66.0
bandit>=1.7.0
